requirements.txt
streamlit>=1.28.0
openai>=1.0.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
python-dotenv>=1.0.0
requests>=2.31.0
fpdf2>=2.7.0
typing-extensions>=4.0.0
watchdog>=3.0.0
